pyproject.toml
[tool.poetry]
name = "rl"
version = "0.1.0"
description = ""
authors = ["Erik Gafni <egafni@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
plotly = ">=5"
pandas = "^1.2.2"
numpy = "^1.18.0"
scipy = "^1.6.0"
requests = "^2.25.1"
matplotlib = "^3.3.4"
seaborn = "^0.11.1"
statsmodels = "^0.12.2"
line-profiler = "^3.2.0"
joblib = "^1.0.1"
tabulate = "^0.8.7"
scikit-learn = "^0.24.1"
tqdm = "^4.56.2"
torch = "^1.8"
gym-super-mario-bros = "^7.3.2"
gym = {extras = ["atari"], version = "^0.18.0"}
tensorboard = "^2.4.1"
opencv-python = "^4.5.1"
simple-parsing = "^0.0.14"
torchvision = "0.9.1"
PyVirtualDisplay = "^2.1"
stable-baselines3 = "^1.0"
dateparser = "^1.0.0"
h5py = "^3"
smart-open = "^5.0.0"
lightning-bolts = "^0.3.3"
fastapi = "^0.64.0"
uvicorn = "^0.13.4"
dash = "^1.20.0"

ray = {extras = ["rllib", "default"], version = "^1.3.0"}
GPUtil = "^1.4.0"
pybullet = "^3.1.7"
ipython-autotime = "^0.3.1"
more-itertools = "^8.8.0"
Glances = "^3.1.7"
dacite = "^1.6.0"
nbval = "^0.9.6"
hyperopt = "^0.2.5"
tensorflow = "^2.5.0"
methodtools = "^0.4.3"
bayesian-optimization = "^1.2.0"
hpbandster = "^0.7.4"
ConfigSpace = "^0.4.19"
nevergrad = "^0.4.3"
pycoingecko = "^2.2.0"
coinmarketcap = "^5.0.3"
jupyter-dash = "^0.4.0"
ipywidgets = "^7.6.5"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
ipywidgets = "^7.6.3"
ipdb = "^0.13.4"
black = "^20.8b1"
pytest = "^6.2.2"
pytest-timeout = "^1.4.2"
pytest-xdist = "^2.2.1"
autoflake = "^1.4"
flake8 = "^3.8.4"
isort = "^5.7.0"
mypy = "^0.800"
pre-commit = "^2.10.1"
jupyterlab = "^3.0.7"
line-profiler = "^3.2.6"
poethepoet = "^0.10.0"
jupyter-contrib-nbextensions = "^0.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | build
  | dist
  | notebooks
  | submodules
  )/
'''

[tool.isort]
profile = "black"
line_length = 120
src_paths = ["oracle", "scripts"]
skip = "submodules"

[tool.pytest.ini_options]
addopts = "--doctest-modules --doctest-glob=\"*.py\" -vv --durations 5 --pdbcls=IPython.terminal.debugger:Pdb --strict-markers"
testpaths = [
    "oracle/tests",
    "notebooks/tutorials"
]
filterwarnings = [
"ignore:.+a deprecated alias for the builtin `.+"
]
[tool.poe.tasks]
force-cuda11 = "python -m pip install torch==1.8.1+cu111 torchvision==0.9.1+cu111 -f https://download.pytorch.org/whl/torch_stable.html"
fix-own="chown -R $UID:$GID"  # specify directory

[tool.mypy]
plugins="numpy.typing.mypy_plugin"
